+++ core/src/main/java/io/snabble/sdk/shoppingcart/ShoppingCartUpdater.kt
@@ -36,6 +36,12 @@ internal class ShoppingCartUpdater(
     private val checkoutApi: DefaultCheckoutApi = DefaultCheckoutApi(project, cart)
     private val handler: Handler = Handler(Looper.getMainLooper())
 
+    /**
+     * Listener that notifies if items are detected that are that cant be found
+     * in the product data base
+     */
+    var onInvalidItemsDetectedListener: ((List<String>) -> Unit)? = null
+
     var lastAvailablePaymentMethods: List<PaymentMethodInfo>? = null
         private set
     var isUpdated = false
@@ -91,6 +97,12 @@ internal class ShoppingCartUpdater(
 
                 override fun onInvalidProducts(products: List<Product>) {
                     cart.invalidProducts = products
+                    error(requestSucceeded = true)
+                }
+
+                override fun onInvalidItems(itemIds: List<String>) {
+                    handler.post { onInvalidItemsDetectedListener?.invoke(itemIds) }
+                    cart.invalidItemIds = itemIds
                     error(requestSucceeded = true)
                 }
 
@@ -161,6 +173,7 @@ internal class ShoppingCartUpdater(
 
         with(cart) {
             invalidProducts = null
+            invalidItemIds = null
             checkLimits()
             notifyPriceUpdate(this)
         }